refactor(tools): give the enforceability judgement its own module
`packages/agent-tools/src/sandbox/workspace-manifest.ts` reached 302 lines
against the 300-line anti-monolith limit, and `pre-push` refused it.
Splitting by responsibility is what that rule asks for, and there is a real
seam here rather than an arbitrary cut: deciding WHETHER a manifest's
controls can be enforced is a different job from applying its entries, and
the applicator was the only reason the two sat together.

`packages/agent-tools/src/sandbox/manifest-enforceability.ts` owns the
judgement, unchanged. The applicator imports it and calls it in the same
place, before any entry is applied.

No behaviour change: tsc 0, the same 15 tests pass, file-size passes.
